Fix authentication and authorization traits.
